Player gun reload time affected by HP
Player guns' reload time is now affected by their HP (less hp = longer reload).
Also optimized corresponding scripts.

+++ Assets/Scripts/CannonManager.cs
@@ -18,7 +18,7 @@ public class CannonManager : MonoBehaviour {
     public TargetManager targetingScript;
 
     public float loadTime = 2f;
-    public int damageOutput = 33;
+    public float damageOutput = 33;
     public float rangeOfGuns = 80f;
 
     private float timeWhenLoaded;
